feat: add NMH file to support 1password integration in Chromium
Main gripe since installing Ungoogled Chromium is that the 1Password
extension's app integration didn't work, even after configuring it as a
custom app. Found a support article that said it's because Ungoogled
Chromium doesn't automatically create the Native Messaging Hosts file
and that it can be created automatically.

This creates that file and enables the integration.